# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Challenge 5 - Cars Express App",
    "version": "3.0.0",
    "description": "Aplikasi penyimpanan data mobil dengan menggunakan expressJS dan database PostgreSQL dengan kebutuhan melakukan CRUD mobil serta mengimplementasikan autentikasi dan otorisasi pengguna.\n\nAvailable Users:\n- SuperAdmin\n  - email: superadmin@gmail.com\n  - password: superadmin123\n- Admin\n  - email: admin123@gmail.com\n  - password: admin123\n- Member\n  - email: user@gmail.com (ex.)\n  - password: userpassword (ex.)",
    "contact": {
      "email": "luthfiyanto1425@gmail.com"
    }
  },
...
  "servers": [
    {
      "url": "localhost:8000/"
    }
  ],
...
  "paths": {
...
    "/cars": {
      "get": {
        "tags": [
          "car"
        ],
        "summary": "Get list of cars",
        "description": "Get list of cars",
        "operationId": "getListCars",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Car"
                }
              }
            }
          },
          "400": {
            "description": "Token invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissingAuthorizationHeader"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/MissingAuthorizationHeader"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorizationError"
                }
              }
            }
          },
          "404": {
            "description": "Car not found or is empty",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoContent"
                }
              }
            }
          },
          "500": {
            "description": "Failed operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "car"
        ],
        "summary": "Create  car by Id",
        "description": "create car by input Id",
        "operationId": "createCarById",
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Car"
                }
              }
            }
          },
          "400": {
            "description": "Token invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissingAuthorizationHeader"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/MissingAuthorizationHeader"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorizationError"
                }
              }
            }
          },
          "403": {
            "description": "Access Denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessDenied"
                }
              }
            }
          },
          "500": {
            "description": "Failed operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
    "/cars/{carId}": {
      "get": {
        "tags": [
          "car"
        ],
        "summary": "Get detail car by Id",
        "description": "Get detail car by input Id",
        "operationId": "getCarById",
        "parameters": [
          {
            "name": "carId",
            "in": "path",
            "description": "ID of the car",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailCar"
                }
              }
            }
          },
          "400": {
            "description": "Token invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissingAuthorizationHeader"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/MissingAuthorizationHeader"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorizationError"
                }
              }
            }
          },
          "404": {
            "description": "Car not found or is empty",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoContent"
                }
              }
            }
          },
          "500": {
            "description": "Failed operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "car"
        ],
        "summary": "Get detail car by Id",
        "description": "Get detail car by input Id",
        "operationId": "updateCarById",
        "parameters": [
          {
            "name": "carId",
            "in": "path",
            "description": "ID of the car",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailCar"
                }
              }
            }
          },
          "400": {
            "description": "Token invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissingAuthorizationHeader"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/MissingAuthorizationHeader"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorizationError"
                }
              }
            }
          },
          "403": {
            "description": "Access Denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessDenied"
                }
              }
            }
          },
          "404": {
            "description": "Car not found or is empty",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoContent"
                }
              }
            }
          },
          "500": {
            "description": "Failed operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "car"
        ],
        "summary": "Delete car by Id",
        "description": "Delete car by input Id",
        "operationId": "deleteCarById",
        "parameters": [
          {
            "name": "carId",
            "in": "path",
            "description": "ID of the car",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedCar"
                }
              }
            }
          },
          "400": {
            "description": "Token invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissingAuthorizationHeader"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/MissingAuthorizationHeader"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorizationError"
                }
              }
            }
          },
          "403": {
            "description": "Access Denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessDenied"
                }
              }
            }
          },
          "404": {
            "description": "Car not found or is empty",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoContent"
                }
              }
            }
          },
          "500": {
            "description": "Failed operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Car": {
        "type": "object",
        "properties": {
          "carId": {
            "type": "string",
            "format": "uuidv4",
            "example": "6e2bc663-5197-441a-957b-bc75e4a2da7c"
          },
          "name": {
            "type": "string",
            "example": "Ford F150"
          },
          "type": {
            "type": "string",
            "example": "small",
            "enum": [
              "small",
              "medium",
              "large"
            ]
          },
          "image": {
            "type": "string",
            "example": "./images/car01.min.jpg"
          },
          "capacity": {
            "type": "integer",
            "example": 2
          },
          "rentPerDay": {
            "type": "integer",
            "example": 200000
          },
          "availableAt": {
            "type": "string",
            "format": "date",
            "example": "2023-03-23T15:49:05.563Z"
          },
          "createdAt": {
            "type": "string",
            "format": "date",
            "example": "2023-03-23T15:49:05.563Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date",
            "example": "2023-03-23T15:49:05.563Z"
          }
        },
        "xml": {
          "name": "car"
        }
      },
      "DetailCar": {
        "type": "object",
        "properties": {
          "carId": {
            "type": "string",
            "format": "uuidv4",
            "example": "6e2bc663-5197-441a-957b-bc75e4a2da7c"
          },
          "name": {
            "type": "string",
            "example": "Ford F150"
          },
          "type": {
            "type": "string",
            "example": "small",
            "enum": [
              "small",
              "medium",
              "large"
            ]
          },
          "image": {
            "type": "string",
            "example": "./images/car01.min.jpg"
          },
          "capacity": {
            "type": "integer",
            "example": 2
          },
          "rentPerDay": {
            "type": "integer",
            "example": 200000
          },
          "availableAt": {
            "type": "string",
            "format": "date",
            "example": "2023-03-23T15:49:05.563Z"
          },
          "createdAt": {
            "type": "string",
            "format": "date",
            "example": "2023-03-23T15:49:05.563Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date",
            "example": "2023-03-23T15:49:05.563Z"
          },
          "createdby": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "example": "7f1e8ec2-e47a-4053-b845-85c8ca895ab0"
              },
              "name": {
                "type": "string",
                "example": "Agus Lapar Buk"
              },
              "email": {
                "type": "string",
                "example": "admin1@email.com"
              },
              "password": {
                "type": "string",
                "example": "adminx11"
              },
              "phoneNumber": {
                "type": "string",
                "example": "089799999999"
              },
              "address": {
                "type": "string",
                "example": "Gunungpati, Semarang"
              },
              "role": {
                "type": "string",
                "example": "ADMIN"
              },
              "createAt": {
                "type": "string",
                "format": "date",
                "example": "2023-10-21T14:39:01.962Z"
              },
              "updateAt": {
                "type": "string",
                "format": "date",
                "example": "2023-10-21T14:39:01.962Z"
              }
            }
          },
          "updatedby": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "example": "7f1e8ec2-e47a-4053-b845-85c8ca895ab0"
              },
              "name": {
                "type": "string",
                "example": "Agus Lapar Buk"
              },
              "email": {
                "type": "string",
                "example": "admin1@email.com"
              },
              "password": {
                "type": "string",
                "example": "adminx11"
              },
              "phoneNumber": {
                "type": "string",
                "example": "089799999999"
              },
              "address": {
                "type": "string",
                "example": "Gunungpati, Semarang"
              },
              "role": {
                "type": "string",
                "example": "ADMIN"
              },
              "createAt": {
                "type": "string",
                "format": "date",
                "example": "2023-10-21T14:39:01.962Z"
              },
              "updateAt": {
                "type": "string",
                "format": "date",
                "example": "2023-10-21T14:39:01.962Z"
              }
            }
          }
        },
        "xml": {
          "name": "updatedcar"
        }
      },
      "DeletedCar": {
        "type": "object",
        "properties": {
          "carId": {
            "type": "string",
            "format": "uuidv4",
            "example": "6e2bc663-5197-441a-957b-bc75e4a2da7c"
          },
          "name": {
            "type": "string",
            "example": "Ford F150"
          },
          "type": {
            "type": "string",
            "example": "small",
            "enum": [
              "small",
              "medium",
              "large"
            ]
          },
          "image": {
            "type": "string",
            "example": "./images/car01.min.jpg"
          },
          "capacity": {
            "type": "integer",
            "example": 2
          },
          "rentPerDay": {
            "type": "integer",
            "example": 200000
          },
          "availableAt": {
            "type": "string",
            "format": "date",
            "example": "2023-03-23T15:49:05.563Z"
          },
          "createdAt": {
            "type": "string",
            "format": "date",
            "example": "2023-03-23T15:49:05.563Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date",
            "example": "2023-03-23T15:49:05.563Z"
          },
          "deletedAt": {
            "type": "string",
            "format": "date",
            "example": "2023-03-23T15:49:05.563Z"
          },
          "createdby": {
            "type": "string",
            "example": "7f1e8ec2-e47a-4053-b845-85c8ca895ab0"
          },
          "updatedby": {
            "type": "string",
            "example": "7f1e8ec2-e47a-4053-b845-85c8ca895ab0"
          },
          "deletedby": {
            "type": "string",
            "example": "7f1e8ec2-e47a-4053-b845-85c8ca895ab0"
          }
        }
      },
...
      "InternalServerError": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "Failed"
          },
          "message": {
            "type": "string",
            "example": "Internal Server Error"
          }
        },
        "xml": {
          "name": "server error"
        }
      },
...
      "AuthorizationError": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "Failed"
          },
          "message": {
            "type": "string",
            "example": "Unauthorized"
          }
        }
      },
      "MissingAuthorizationHeader": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "Failed"
          },
          "message": {
            "type": "string",
            "example": "Authorization Header required"
          }
        }
      },
      "AccessDenied": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "Failed"
          },
          "message": {
            "type": "string",
            "example": "Forbidden Access"
          }
        }
      },
      "NoContent": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "Failed"
          },
          "message": {
            "type": "string",
            "example": "Not Found"
          }
        }
      }
    }
  }
}